pyproject.toml
[project]
name = "sututu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "akshare-mcp>=0.1.3",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "jinja2>=3.1.2",
    "python-multipart>=0.0.6",
]
